+++ src/Exceptions/ResponseException.php
@@ -13,7 +13,6 @@ namespace ShugaChara\Framework\Exceptions;
 
 use Exception;
 use ShugaChara\Framework\Contracts\ThrowExceptionReturnInterface;
-use ShugaChara\Framework\Helpers\FHelper;
 
 /**
  * 用于接口response响应错误
@@ -30,7 +29,7 @@ class ResponseException implements ThrowExceptionReturnInterface
     {
         // TODO: Implement getReturn() method.
 
-        return (FHelper::c()->get('is_debug') == 'true') ?
+        return (fn()->c()->get('is_debug') == 'true') ?
             DebugLogsException::getReturn($exception)
             : null;
     }