+++ src/Swoole/Server.php
@@ -1,6 +1,6 @@
 <?php
 // +----------------------------------------------------------------------
-// | Created by linshan. 版权所有 @
+// | Created by ShugaChara. 版权所有 @
 // +----------------------------------------------------------------------
 // | Copyright (c) 2020 All rights reserved.
 // +----------------------------------------------------------------------
@@ -9,18 +9,23 @@
 // | Author: kaka梦很美 <1099013371@qq.com>
 // +----------------------------------------------------------------------
 
+/*
+|--------------------------------------------------------------------------
+| shugachara Swoole Server
+|--------------------------------------------------------------------------
+ */
+
 namespace ShugaChara\Framework\Swoole;
 
 use ShugaChara\Framework\Contracts\PoolInterface;
-use ShugaChara\Framework\Helpers\FHelper;
 use ShugaChara\Http\SwooleServerRequest;
 use ShugaChara\Swoole\EventsRegister;
-use ShugaChara\Swoole\Server as SwooleServer;
 use ShugaChara\Swoole\Manager\Timer;
+use ShugaChara\Swoole\Server as SwooleServer;
 use ShugaChara\Swoole\SwooleHelper;
-use swoole_server;
 use swoole_http_request;
 use swoole_http_response;
+use swoole_server;
 use swoole_websocket_server;
 
 /**
@@ -30,21 +35,21 @@ use swoole_websocket_server;
 class Server extends SwooleServer
 {
     /**
-     * 注册默认回调
+     * Register the default callback
      * @param swoole_server $server
      * @param               $server_name
      */
     public function registerDefaultCallback(swoole_server $server, $server_name)
     {
-        if (in_array($server_name, [SwooleServer::SWOOLE_HTTP_SERVER, SwooleServer::SWOOLE_WEBSOCKET_SERVER])) {
-            // 注册请求事件 request event
+        if (in_array($server_name, [static::SWOOLE_HTTP_SERVER, static::SWOOLE_WEBSOCKET_SERVER])) {
+            // Registration request event
             $this->getEventsRegister()->on(
                 $server,
                 EventsRegister::onRequest,
                 function (swoole_http_request $swooleRequest, swoole_http_response $swooleResponse) use ($server) {
-                    // 转接 Swoole request 对象
+                    // Transfer Swoole request object
                     $request = SwooleServerRequest::createServerRequestFromSwoole($swooleRequest);
-                    $response = FHelper::app()->handleRequest($request);
+                    $response = fn()->app()->handleRequest($request);
                     foreach ($response->getHeaders() as $key => $header) {
                         $swooleResponse->header($key, $response->getHeaderLine($key));
                     }
@@ -66,22 +71,22 @@ class Server extends SwooleServer
                 }
             );
 
-            // 注册默认的 websocket onMessage event
-            if ($server_name == SwooleServer::SWOOLE_WEBSOCKET_SERVER) {
+            // Register the default websocket onMessage event
+            if ($server_name == static::SWOOLE_WEBSOCKET_SERVER) {
                 $this->getEventsRegister()->addEvent('message', function(swoole_websocket_server $server, $frame) {});
             }
 
-            // 注册默认的 worker start event
+            // Register the default worker start event
             $this->getEventsRegister()->addEvent(
                 EventsRegister::onWorkerStart,
                 function (swoole_server $server, $workerId) use ($server_name) {
                     if(PHP_OS != 'Darwin'){
-                        if( ($workerId < FHelper::c()->get('swoole.' . $server_name . '.setting.worker_num')) && $workerId >= 0){
+                        if( ($workerId < fn()->c()->get('swoole.' . $server_name . '.setting.worker_num')) && $workerId >= 0){
                             SwooleHelper::setProcessRename(("{$server_name}.Worker.{$workerId}"));
                         }
                     }
 
-                    // 建立Pool连接池
+                    // Establish Pool connection pool
                     foreach (container()->getContainerServices() as $service) {
                         if ($service instanceof PoolInterface) {
                             $service->initPool();
@@ -90,18 +95,19 @@ class Server extends SwooleServer
                 }
             );
 
-            // 注册默认的 worker exit event
+            // Register the default worker exit event
             $this->getEventsRegister()->addEvent(
                 EventsRegister::onWorkerExit,
                 function () {
-                    Timer::getInstance()->clearAll();
+                    Timer::clearAll();
                 }
             );
 
-            // 注册onTask事件
+            // Register onTask event
             $this->getEventsRegister()->addEvent(EventsRegister::onTask, function (swoole_server $serv, int $task_id, int $src_worker_id, mixed $data) {
-                // ...待填充
+                // ...To be filled
             });
         }
     }
 }
+