+++ src/Traits/Application.php
@@ -11,9 +11,6 @@
 
 namespace ShugaChara\Framework\Traits;
 
-use ShugaChara\Core\Utils\Helper\ArrayHelper;
-use ShugaChara\Framework\Components\Alias;
-
 /**
  * Trait Application
  * @package ShugaChara\Framework\Traits
@@ -21,7 +18,51 @@ use ShugaChara\Framework\Components\Alias;
 trait Application
 {
     /**
-     * 获取应用根目录
+     * Application name
+     * @return mixed|null
+     */
+    public function getName()
+    {
+        return fn()->c()->get('app_name', '');
+    }
+
+    /**
+     * Application version
+     * @return mixed|null
+     */
+    public function getVersion()
+    {
+        return fn()->c()->get('app_version', 'v1.0.0');
+    }
+
+    /**
+     * Application logo
+     * @return string
+     */
+    public function getLogo()
+    {
+        return <<<LOGO
+                                                                                                
+ ┌───┐   ┌───┬───┬───┬───┐ ┌───┬───┬───┬───┐ ┌───┬───┬───┬───┐ ┌───┬───┬───┐      Server                
+ │Esc│   │ F1│ F2│ F3│ F4│ │ F5│ F6│ F7│ F8│ │ F9│F10│F11│F12│ │P/S│S L│P/B│  ┌┐    ┌┐    ┌┐               
+ └───┘   └───┴───┴───┴───┘ └───┴───┴───┴───┘ └───┴───┴───┴───┘ └───┴───┴───┘  └┘    └┘    └┘               
+ ┌───┬───┬───┬───┬───┬───┬───┬───┬───┬───┬───┬───┬───┬───────┐ ┌───┬───┬───┐ ┌───┬───┬───┬───┐             
+ │~ `│! 1│@ 2│# 3│$ 4│% 5│^ 6│& 7│* 8│( 9│) 0│_ -│+ =│ BacSp │ │Ins│Hom│PUp│ │N L│ / │ * │ - │             
+ ├───┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─────┤ ├───┼───┼───┤ ├───┼───┼───┼───┤             
+ │ Tab │ Q │ W │ E │ R │ T │ Y │ U │ I │ O │ P │{ [│} ]│ | \ │ │Del│End│PDn│ │ 7 │ 8 │ 9 │   │             
+ ├─────┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴─────┤ └───┴───┴───┘ ├───┼───┼───┤ + │             
+ │ Caps │ A │ S │ D │ F │ G │ H │ J │ K │ L │: ;│" '│ Enter  │               │ 4 │ 5 │ 6 │   │
+ ├──────┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴────────┤     ┌───┐     ├───┼───┼───┼───┤
+ │ Shift  │ Z │ X │ C │ V │ B │ N │ M │< ,│> .│? /│  Shift   │     │ ↑ │     │ 1 │ 2 │ 3 │   │
+ ├─────┬──┴─┬─┴──┬┴───┴───┴───┴───┴───┴──┬┴───┼───┴┬────┬────┤ ┌───┼───┼───┐ ├───┴───┼───┤ E││
+ │ Ctrl│    │Alt │         Space         │ Alt│    │    │Ctrl│ │ ← │ ↓ │ → │ │   0   │ . │←─┘│
+ └─────┴────┴────┴───────────────────────┴────┴────┴────┴────┘ └───┴───┴───┘ └───────┴───┴───┘
+
+LOGO;
+    }
+
+    /**
+     * Get the application root directory
      * @return mixed
      */
     public function getRootDirectory()
@@ -30,38 +71,10 @@ trait Application
     }
 
     /**
-     * 设置命令行参数
-     * @param $argv
-     */
-    protected function setArgv($argv)
-    {
-        $this->argv['exec'] = ArrayHelper::get($argv, 0);
-        $this->argv = [
-            'exec'  =>  ArrayHelper::get($argv, 0),
-            'name'  =>  ArrayHelper::get($argv, 1),
-        ];
-
-        unset($argv[0], $argv[1]);
-
-        $this->argv['param'] = array_values($argv);
-
-        Alias::set('argv', $this->argv);
-    }
-
-    /**
-     * 获取命令行参数
-     * @return mixed
-     */
-    public function getArgv()
-    {
-        return Alias::get('argv');
-    }
-
-    /**
-     * 服务容器注册
+     * Service container registration
      * @param array $services
      */
-    public function serviceProviderRegister(array $services)
+    public function registerServiceProviders(array $services)
     {
         foreach ($services as $service) {
             (new $service)->register(container());
@@ -69,7 +82,7 @@ trait Application
     }
 
     /**
-     * 应用是否已运行
+     * Whether the app is already running
      * @return bool
      */
     protected function isExecute(): bool