+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 2019.06.06 `v1.0.0`
 
-- 基于 Swoole 扩展开发的一套高性能 WebAPI 框架 (支持 php-fpm / swoole 模式)
+- 基于 Swoole 扩展开发的一套高性能 WebApi 框架 (支持 php-fpm / swoole 模式)